+++ src/NNHelferlein.jl
@@ -65,6 +65,7 @@ export AbstractNN, Classifier, Regressor, Chain, VAE,          # chains
        Linear, GaussianNoise,
        GlobalAveragePooling, 
        Activation, Logistic, Sigm, Relu,
+       ResNetBlock,
        get_hidden_states, get_cell_states,
        set_hidden_states!, set_cell_states!,
        reset_hidden_states!, reset_cell_states!,